Refactor components related to chatflow configuration
