+++ packages/ui/src/ui-component/extended/AllowedDomains.jsx
